text-image/requirements.txt
fastapi==0.115.12
uvicorn==0.34.3
google-cloud-aiplatform==1.67.1
google-genai==1.20.0
google-auth==2.27.0
pillow==11.2.1
python-dotenv
motor==3.3.2
pymongo==4.6.1